minor cleanup, add encoding test, and add ability to turn off metrics logging
