+++ src/deadline_worker_agent/startup/config_file.py
@@ -39,6 +39,7 @@ class LoggingConfigSection(BaseModel):
     verbose: Optional[bool] = None
     worker_logs_dir: Optional[Path] = None
     local_session_logs: Optional[bool] = None
+    host_metrics_logging: Optional[bool] = None
     host_metrics_logging_interval_seconds: Optional[float] = None
 
 
@@ -108,6 +109,8 @@ class ConfigFile(BaseModel):
             output_settings["worker_logs_dir"] = self.logging.worker_logs_dir
         if self.logging.local_session_logs is not None:
             output_settings["local_session_logs"] = self.logging.local_session_logs
+        if self.logging.host_metrics_logging is not None:
+            output_settings["host_metrics_logging"] = self.logging.host_metrics_logging
         if self.logging.host_metrics_logging_interval_seconds is not None:
             output_settings[
                 "host_metrics_logging_interval_seconds"
