+++ src/deadline_worker_agent/worker.py
@@ -6,21 +6,21 @@ import json
 import signal
 import sys
 import traceback
+from contextlib import nullcontext
 from concurrent.futures import Executor, Future, ThreadPoolExecutor, wait
 from datetime import datetime, timedelta, timezone
-from logging import getLogger, Logger
-from threading import Event, Timer
+from logging import getLogger
+from threading import Event
 from types import FrameType
 from typing import Any, NamedTuple, cast
 from pathlib import Path
 
 import boto3
-import os
 import requests
-import psutil
 
 from .boto import DeadlineClient
 from .errors import ServiceShutdown
+from .metrics import HostMetricsLogger
 from .scheduler import WorkerScheduler
 from .sessions import Session
 from .startup.config import ImpersonationOverrides
@@ -70,7 +70,7 @@ class Worker:
     _logs_client: boto3.client
     _boto_session: WorkerBoto3Session
     _worker_persistence_dir: Path
-    _host_metrics_logging_interval_seconds: float
+    _host_metrics_logger: HostMetricsLogger | None = None
 
     def __init__(
         self,
@@ -86,7 +86,8 @@ class Worker:
         cleanup_session_user_processes: bool,
         worker_persistence_dir: Path,
         worker_logs_dir: Path | None,
-        host_metrics_logging_interval_seconds: float,
+        host_metrics_logging: bool,
+        host_metrics_logging_interval_seconds: float | None,
     ) -> None:
         self._deadline_client = deadline_client
         self._s3_client = s3_client
@@ -109,7 +110,14 @@ class Worker:
         self._stop = Event()
         self._boto_session = boto_session
         self._worker_persistence_dir = worker_persistence_dir
-        self._host_metrics_logging_interval_seconds = host_metrics_logging_interval_seconds
+
+        if host_metrics_logging:
+            assert (
+                host_metrics_logging_interval_seconds is not None
+            ), "host_metrics_logging_interval_seconds is required if host metrics logging is enabled"
+            self._host_metrics_logger = HostMetricsLogger(
+                logger=logger, interval_s=host_metrics_logging_interval_seconds
+            )
 
         signal.signal(signal.SIGTERM, self._signal_handler)
         signal.signal(signal.SIGINT, self._signal_handler)
@@ -174,10 +182,7 @@ class Worker:
                 session=self._boto_session,
                 failure_callback=self._aws_credentials_refresh_failure,
             ),
-            HostMetricsLogger(
-                logger=logger,
-                interval_s=self._host_metrics_logging_interval_seconds,
-            ),
+            self._host_metrics_logger or nullcontext(),
         ):
             scheduler_future = self._executor.submit(self._scheduler.run)
             futures: list[Future[Any]] = [
@@ -423,96 +428,3 @@ class Worker:
         if response.status_code == 200:
             return response.text == "Terminated"
         return False
-
-
-class HostMetricsLogger:
-    """Context manager that regularly logs host metrics"""
-
-    logger: Logger
-    interval_s: float
-    _timer: Timer | None
-    _prev_network: Any | None
-
-    def __init__(self, logger: Logger, interval_s: float) -> None:
-        assert interval_s > 0, "interval_s must be a positive number"
-        self._timer = None
-        self._prev_network = None
-        self.logger = logger
-        self.interval_s = interval_s
-
-    def __enter__(self) -> HostMetricsLogger:
-        self.log_metrics()
-        return self
-
-    def __exit__(self, type, value, traceback) -> None:
-        if self._timer:
-            self._timer.cancel()
-            self._timer = None
-
-    def log_metrics(self):
-        """
-        Queries information about the host machine and logs the information as a space-delimited
-        line of the form: <label> <value> ...
-        """
-        memory = psutil.virtual_memory()
-        swap = psutil.swap_memory()
-        disk = psutil.disk_usage(os.sep)
-
-        # On Windows it may be necessary to issue diskperf -y command from cmd.exe first in order to enable IO counters
-        disk_counters = psutil.disk_io_counters(nowrap=True)
-        if disk_counters is None:
-            disk_read = disk_write = "NOT_AVAILABLE"
-        elif not (hasattr(disk_counters, "read_bytes") and hasattr(disk_counters, "write_bytes")):
-            # TODO: Support disk speed on NetBSD and OpenBSD
-            disk_read = disk_write = "NOT_SUPPORTED"
-        else:
-            disk_read = str(round(disk_counters.read_bytes / self.interval_s))
-            disk_write = str(round(disk_counters.write_bytes / self.interval_s))
-
-        # We need to poll network IO to get rate
-        network = psutil.net_io_counters(nowrap=True)
-        if network is None:
-            network_sent = network_recv = "NOT_AVAILABLE"
-        else:
-            if self._prev_network:
-                network_sent_bps = round(
-                    (network.bytes_sent - self._prev_network.bytes_sent) / self.interval_s
-                )
-                network_recv_bps = round(
-                    (network.bytes_recv - self._prev_network.bytes_recv) / self.interval_s
-                )
-            else:
-                network_sent_bps = network_recv_bps = 0
-            network_sent = str(network_sent_bps)
-            network_recv = str(network_recv_bps)
-        self._prev_network = network
-
-        stats = {
-            "cpu-usage-percent": str(psutil.cpu_percent()),
-            "memory-total-bytes": str(memory.total),
-            "memory-used-bytes": str(memory.total - memory.available),
-            "memory-used-percent": str(memory.percent),
-            "swap-used-bytes": str(swap.used),
-            "total-disk-bytes": str(disk.total),
-            "total-disk-used-bytes": str(disk.used),
-            "total-disk-used-percent": str(round(disk.used / disk.total, ndigits=1)),
-            "user-disk-available-bytes": str(disk.free),
-            "network-sent-bytes-per-second": network_sent,
-            "network-recv-bytes-per-second": network_recv,
-            "disk-read-bytes-per-second": disk_read,
-            "disk-write-bytes-per-second": disk_write,
-        }
-
-        # Output as space-delimited "key value" pairs for consumption by Cloudwatch to use as metrics
-        self.logger.info(" ".join(" ".join(kvp) for kvp in stats.items()))
-        self._set_timer()
-
-    def _set_timer(self) -> None:
-        """
-        Sets the timer to log the host metrics at a regular interval.
-
-        Args:
-            interval_s (float): The interval in seconds to print the host metrics at.
-        """
-        timer = Timer(self.interval_s, self.log_metrics)
-        timer.start()
