+++ test/unit/startup/test_settings.py
@@ -123,6 +123,13 @@ FIELD_TEST_CASES: list[FieldTestCaseParams] = [
         expected_default_factory_return_value=None,
     ),
     FieldTestCaseParams(
+        field_name="host_metrics_logging",
+        expected_type=bool,
+        expected_required=False,
+        expected_default=True,
+        expected_default_factory_return_value=None,
+    ),
+    FieldTestCaseParams(
         field_name="host_metrics_logging_interval_seconds",
         expected_type=float,
         expected_required=False,
